requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-magic-bin==0.4.14
Pillow==9.5.0
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.2.1 